Expose range on map
