+++ packages/storage-plus/src/lib.rs
@@ -1,3 +1,4 @@
+mod keys;
 mod length_prefixed;
 mod map;
 mod namespace_helpers;
@@ -5,15 +6,8 @@ mod path;
 mod prefix;
 mod type_helpers;
 
+pub use keys::{Prefixer, PrimaryKey};
 pub use map::Map;
 pub use path::Path;
 #[cfg(feature = "iterator")]
 pub use prefix::Prefix;
-
-#[cfg(test)]
-mod test {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
