+++ packages/storage-plus/src/prefix.rs
@@ -1,5 +1,4 @@
 #![cfg(feature = "iterator")]
-
 use serde::de::DeserializeOwned;
 use serde::Serialize;
 use std::marker::PhantomData;
